Extract error classes into a separate file
